build: add build-schema script to build schemas
